pyproject.toml
[tool.poetry]
name = "parsermp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
pytest-playwright = "^0.5.0"
loguru = "^0.7.2"
html-to-json = "^2.0.0"
lxml = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
